requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
openai
chromadb
numpy
pydantic
sentence-transformers
pytest